+++ src/library/DIPS.Mobile.UI/Components/Counters/Counter.Properties.cs
@@ -48,12 +48,33 @@ public partial class Counter
         typeof(Counter),
         propertyChanged: (bindable, _, _) => ((Counter)bindable).OnIsSecondaryErrorChanged());
 
+    public static readonly BindableProperty IsFlippedProperty = BindableProperty.Create(
+        nameof(IsFlipped),
+        typeof(bool),
+        typeof(Counter),
+        propertyChanged: (bindable, _, _) => ((Counter)bindable).OnIsFlippedChanged());
+
+    /// <summary>
+    /// Determines whether the counter is flipped, meaning the primary value is displayed on the right side and the secondary value on the left side.
+    /// </summary>
+    public bool IsFlipped
+    {
+        get => (bool)GetValue(IsFlippedProperty);
+        set => SetValue(IsFlippedProperty, value);
+    }
+    
+    /// <summary>
+    /// Determines whether the secondary value is in an error state and should display an error icon.
+    /// </summary>
     public bool IsSecondaryError
     {
         get => (bool)GetValue(IsSecondaryErrorProperty);
         set => SetValue(IsSecondaryErrorProperty, value);
     }    
     
+    /// <summary>
+    /// Determines whether the primary value is in an error state and should display an error icon.
+    /// </summary>
     public bool IsError
     {
         get => (bool)GetValue(IsErrorProperty);