+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [47.9.0]
-- [Counter] Add `IsError` and `IsSecondaryError` properties
+- [Counter] Add `IsError`, `IsSecondaryError` and `IsFlipped` properties
 
 ## [47.8.1] 
 - [TabView] Small style fixes